requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
cffi==1.17.1
click==8.2.0
cryptography==44.0.3
ecdsa==0.19.1
fastapi==0.115.12
freezegun==1.2.2
greenlet==3.2.2
h11==0.16.0
httptools==0.6.4
httpx==0.25.2
idna==3.10
openpyxl==3.1.2
passlib==1.7.4
Pillow==10.1.0
pyasn1==0.4.8
pycparser==2.22
pydantic[email]==2.11.4
pydantic_core==2.33.2
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
pytest-mock==3.12.0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
openai==1.78.1
